Started implementing movements in the board.
I think the smartest thing is to implement a function that checks all legal moves of a piece. Then, when performing a move, we can first check if it is in the list of legal moves.

With that function, we can also get all legals moves by calling the first function on every square