Changed the test code to play against AlphaZero
Added arguments in the config file to control the percentage of early stopping and the limit value to stop

Started to stop early games with -0.5 instead -0.05. The RL needs bad demonstrations as well, and filling the Replay Buffer with only one type of demos, the RL train can become biased. (The early stopping are the first demonstrations to hit the buffer)Â